Move fly/traits to fly/concepts
Now that the traits header is removed, it makes more sense this way.
