+++ test/concepts/concepts.cpp
@@ -1,4 +1,4 @@
-#include "fly/traits/concepts.hpp"
+#include "fly/concepts/concepts.hpp"
 
 #include "catch2/catch_test_macros.hpp"
