+++ test/types/json/json_helpers.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "fly/traits/concepts.hpp"
+#include "fly/concepts/concepts.hpp"
 #include "fly/types/json/json.hpp"
 #include "fly/types/json/json_exception.hpp"
 #include "fly/types/string/string.hpp"
